// frontend/src/views/affect/AF/AWAF00101G/type.ts
/** ----------------------------------------------------------------
 * 業務名称　: 養鶏-互助防疫システム
 * 機能概要　: ログイン
 * 　　　　　  インターフェース定義
 * 作成日　　: 2023.08.30
 * 作成者　　: 屠
 * 変更履歴　:
 * -----------------------------------------------------------------*/
import { Enumログイン区分 } from '#/Enums'

//-------------------------------------------------------------------
//リクエスト
//-------------------------------------------------------------------
/** ログイン処理 */
export interface LoginRequest extends DaRequestBase {
  /** パスワード */
  pword: string
  /** ログイン区分 */
  kbn: Enumログイン区分
}

//-------------------------------------------------------------------
//レスポンス
//-------------------------------------------------------------------
/** ログイン処理(成功) */
export interface LoginResponse extends DaResponseBase {
  /** トークン(ベースロジック) */
  token?: string
  /** ユーザー情報(共通) */
  userinfo?: UserInfoVM
  /** 支所リスト(登録支所選択用) */
  sisyolist: CmSisyoVM[]
  /** 警告フラグ */
  pwdmsgflg?: boolean
}

//-------------------------------------------------------------------
//ビューモデル
//-------------------------------------------------------------------
/** ユーザー共通 */
export interface UserInfoVM {
  /** ユーザID */
  userid: string
  /** ユーザー名 */
  usernm: string
  /** 所属名 */
  syozokunm: string
  /** 管理者フラグ */
  kanrisyaflg: boolean
}